moving tag routes to separate blueprint and file
