+++ app/assess/templates/change_tags.html
@@ -18,7 +18,7 @@
                     <div class="govuk-hint govuk-!-margin-bottom-6">
                         Select the tags you want to apply to this assessment
                     </div>
-                    <a href="{{ url_for('assess_bp.create_tag', fund_id=state.fund_id, round_id=state.round_id) }}"
+                    <a href="{{ url_for('tag_bp.create_tag', fund_id=state.fund_id, round_id=state.round_id) }}"
                        role="button"
                        class="govuk-button govuk-button--secondary">
                         Create new tag
