+++ app/assess/templates/macros/fund_dashboard_summary.html
@@ -9,7 +9,7 @@
         <ul class="govuk-summary-card__actions">
             <li class="govuk-summary-card__action">
                 <a class="govuk-link govuk-link__white_font govuk-!-margin-right-4"
-                   href="{{ url_for('assess_bp.load_fund_round_tags', fund_id=fund_id, round_id=round_id) }}">
+                   href="{{ url_for('tag_bp.load_fund_round_tags', fund_id=fund_id, round_id=round_id) }}">
                     Manage tags
                 </a>
             </li>
