+++ app/assess/templates/manage_tags.html
@@ -17,7 +17,7 @@
 
         <div class="govuk-grid-row govuk-!-text-align-left">
             <div class="govuk-grid-column-full">
-                <a href="{{ url_for('assess_bp.create_tag', fund_id = fund_round['fund_id'], round_id = fund_round['round_id']) }}"
+                <a href="{{ url_for('tag_bp.create_tag', fund_id = fund_round['fund_id'], round_id = fund_round['round_id']) }}"
                     class="govuk-button govuk-button--secondary" data-module="govuk-button">
                     Create new tag
                 </a>
