+++ requirements-dev.txt
@@ -168,10 +168,6 @@ gitpython==3.1.31
     # via bandit
 govuk-frontend-jinja==2.5.0
     # via -r requirements.txt
-greenlet==2.0.2
-    # via
-    #   -r requirements.txt
-    #   sqlalchemy
 gunicorn==20.1.0
     # via
     #   -r requirements.txt
