+++ requirements.txt
@@ -93,8 +93,6 @@ funding-service-design-utils==2.0.13
     # via -r requirements.in
 govuk-frontend-jinja==2.5.0
     # via -r requirements.in
-greenlet==2.0.2
-    # via sqlalchemy
 gunicorn==20.1.0
     # via funding-service-design-utils
 idna==3.4
